Events funciona

+++ app/controllers/events_controller.rb
@@ -1,8 +1,16 @@
+module EventsHelper
+   def same_user_as_current(user_id)
+      return user_id == current_user.id 
+   end
+end
+
 class EventsController < ApplicationController    
  #   before_action :find_event, only: [:show, :edit, :update, :destroy]       
     before_action :authenticate_user 
-    before_action :event_auth, only: [:edit, :update, :destroy]    
+  #  before_action :event_auth, only: [:edit, :update, :destroy]    
     include UtilsModule
+    include EventsHelper
+    
     
         def index    
             events = Event.all.order(created_at: :desc)    
@@ -40,22 +48,33 @@ class EventsController < ApplicationController
 
         def update    
             event = Event.find_by_id(params[:event_id])
-            if event.nil?
-                sendStatus("Event does not exist", :not_found)
-            else
-               if event.update(event_params) and event.location.update(location_params)
-                    render json: { event: event, location: Location.find_by_id(event.location_id)}, status: :ok
+            if same_user_as_current(event.user_id)
+                if event.nil?
+                    sendStatus("Event does not exist", :not_found)
                 else
-                    sendStatus("Error modifying event", :conflict, event.errors) 
+                    l = Location.find_by_id(event.location_id).update(location_params)
+                    e = event.update(event_params) 
+                  #  if e.update(event_params) and l.update(location_params)
+                        render json: { event: event, location: Location.find_by_id(event.location_id)}, status: :ok
+                 #   else
+                #       sendStatus("Error modifying event", :conflict, e.errors) 
+                   # end
                 end
+            else
+                sendStatus("Current user != creator of the event", :conflict, event.errors) 
             end
+            
         end
                 
 
         def destroy    
             event = Event.find_by_id(params[:event_id])
-            event.destroy    
-            render json: @event, status: :destroy      
+            if same_user_as_current(event.user_id)
+                event.destroy    
+                render json: event, status: :destroy 
+            else
+                sendStatus("Current user != creator of the event", :conflict, event.errors) 
+            end
         end    
 
         private    
@@ -73,8 +92,8 @@ class EventsController < ApplicationController
         end   
 
         def event_auth    
-            if current_user != event.user    
-               redirect_to(root_path)    
+            if current_user.id != event.user_id    
+               sendStatus("The user is not the creator of the event", :conflict)  
             end    
         end    
-    
+end
